requirements.txt
numpy
pandas
scikit-learn
streamlit
python-dotenv
groq
